refactor search meal use case
