+++ src/main/kotlin/logic/MealRepository.kt
@@ -4,5 +4,4 @@ import org.example.model.Meal
 
 interface MealRepository {
     fun getAllMeals(): List<Meal>
-    fun searchMealsByCountry(country: String): List<Meal>
 }
